Move `topology::build_or_log_errors` into `TopologyPieces`
