+++ src/topology/test/mod.rs
@@ -19,7 +19,7 @@ use crate::{
         },
         start_topology, trace_init,
     },
-    topology::{self, RunningTopology, TopologyPieces},
+    topology::{RunningTopology, TopologyPieces},
 };
 use futures::{future, stream, StreamExt};
 use tokio::{
@@ -785,7 +785,7 @@ async fn topology_required_healthcheck_fails_start() {
     let mut config = basic_config_with_sink_failing_healthcheck();
     config.healthchecks.require_healthy = true;
     let diff = ConfigDiff::initial(&config);
-    let pieces = topology::build_or_log_errors(&config, &diff, HashMap::new())
+    let pieces = TopologyPieces::build_or_log_errors(&config, &diff, HashMap::new())
         .await
         .unwrap();
 
@@ -798,7 +798,7 @@ async fn topology_required_healthcheck_fails_start() {
 async fn topology_optional_healthcheck_does_not_fail_start() {
     let config = basic_config_with_sink_failing_healthcheck();
     let diff = ConfigDiff::initial(&config);
-    let pieces = topology::build_or_log_errors(&config, &diff, HashMap::new())
+    let pieces = TopologyPieces::build_or_log_errors(&config, &diff, HashMap::new())
         .await
         .unwrap();
     assert!(RunningTopology::start_validated(config, diff, pieces)
